feat(supabase): add Supabase course repository implementations for browser and server
- Introduce SupabaseCourseRepositoryBrowser and SupabaseCourseRepositoryServer classes
- Implement getCourses method with filtering, searching, pagination
- Add CourseRepository interface and CourseFilters type
- Refactor getCourses service to use repository abstraction
- Add factory function to select repository based on environment
- Improve supabase client creation with mock support and clearer error
- Minor fixes in dashboard service and supabase client error message
